Move empty_map_widget.dart to test package
