+++ example/integration_test/empty_map_widget.dart
